Moved NewEncoder method in to a more fiting, restricting to single name
